Gallery/Images V1
Added fancy box which allows for the addition of multiple images. I have used this for the gallery at the moment but will possibly use this for the Chalets part as these are the main images. Added some more styles to improve how the images sit ont he page as i do not want them sitting a the edge of the page