FinalV1
Changed out the contact page + changed alot of the widths and colours. Fixed up small problems. Removed some temp text and replaced with more usefull info. Fixed the email button to link to gmail. Fixed the buttons not working correctly when zoombing in an out(overlay). added buttons to link to chalet headings. still need to fix the alt texts though. added a viewport as i was given feedback to do so.